Created MainMenuScene and added background
