+++ Assets/DataAssets/Materials/DarkStormSkybox.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25f88e5cbb6683249af2f4586b767fec
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
