+++ Assets/DataAssets/Materials/DarkStormSkybox/DarkStorm.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2039a0ad0a9d05549928b70362f54902
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
