+++ Assets/DataAssets/Materials/DarkStormSkybox/FreebieHdri.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d78fe5a22bec38d43bbe7e846ac3925a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
