+++ Assets/DataAssets/Materials/DarkStormSkybox/Hdri setup.txt
@@ -1,0 +1,12 @@
+to use the skyboxs simply drag-and-drop into your scene or open the lighting tab and replace the skybox material.
+Things you might want to adjust for your scene is the intensity and the exposure.
+I also recommend going to edit>project settings>player tab and changing the color space to Linear for better light
+accuracy.
+
+On the occasion of a bad import and skyboxs are dissociated:
+	Select the hdr and in the inspector change Texture Shape to Cube
+	Create new Material and change it to Skybox/Cubemap
+	Then drag and drop into the scene
+
+
+For Mobile switch the Six-sided Skybox's shader from Skybox/6 Sided to Moblie/Skybox
