+++ Assets/DataAssets/Materials/DarkStormSkybox/Hdri setup.txt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: add36ebe5ee5af347a876cdf41d00709
+timeCreated: 1509147461
+licenseType: Store
+TextScriptImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
