+++ Assets/Scripts/Menus/PauseMenu.cs
@@ -1,0 +1,65 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class PauseMenu : MonoBehaviour
+{
+    public GameObject UIMenu;
+
+    private bool GameIsPaused
+    {
+        get
+        {
+            return GameManager.gameIsPaused;
+        }
+        set
+        {
+            GameManager.gameIsPaused = value;
+        }
+    }
+
+    void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.Escape))
+        {
+            if (GameIsPaused)
+            {
+                Resume();
+            }
+            else
+            {
+                Pause();
+            }
+        }
+    }
+    private void Pause()
+    {
+        UIMenu.SetActive(true);
+        Time.timeScale = 0f;
+        GameIsPaused = true;
+    }
+
+    public void Resume()
+    {
+        UIMenu.SetActive(false);
+        Time.timeScale = 1f;
+        GameIsPaused = false;
+    }
+
+    public void Restart()
+    {
+
+    }
+
+    public void LoadMenu()
+    {
+
+    }
+
+    public void QuitGame()
+    {
+
+    }
+}
