--- Assets/Scripts/PauseMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bc2f8f476114b8649b821e8119b29b14
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
